requirements.txt
jass-kit

# unittests
pytest